Update on "migrate matrix_exp to opInfo tests"
Migrating `matrix_exp` to OpInfo testing.

It doesn't have an `out` variant, to I explicitly skipped `out` testing.


Differential Revision: [D27628966](https://our.internmc.facebook.com/intern/diff/D27628966)

[ghstack-poisoned]

+++ torch/testing/_internal/common_methods_invocations.py
@@ -3221,7 +3221,7 @@ op_db: List[OpInfo] = [
            dtypesIfCUDA=all_types_and_complex_and(torch.bool, torch.half, torch.bfloat16),
            sample_inputs_func=sample_inputs_masked_select),
     OpInfo('matrix_exp',
-           dtypesIfCPU=floating_and_complex_types_and(torch.float16),
+           dtypesIfCPU=floating_and_complex_types_and(torch.bfloat16),
            dtypesIfCUDA=floating_and_complex_types_and(torch.float16),
            sample_inputs_func=sample_inputs_matrix_exp,
            supports_out=False),
